fix before demo

+++ backend/src/main/java/com/HTPj/htpj/service/impl/UserServiceImpl.java
@@ -115,6 +115,9 @@ public class UserServiceImpl implements UserService {
 
     @PreAuthorize("hasRole('ADMIN')")
     public void deleteUser(String userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new AppException(ErrorCode.USER_NOT_EXISTED);
+        }
         userRepository.deleteById(userId);
     }
